requirements.txt
torch
torchvision
numpy
opencv-python
pillow
scikit-learn
matplotlib
tqdm
